requirements.txt
pandas
bs4
numpy
pyarrow
fastparquet
requests
pymongo
bs4
psycopg2-binary
dotenv
sentence-transformers